+++ R/make_wordcloud.R
@@ -1,74 +1,74 @@
-#' Create a word cloud from a frequency table of words
+#' Create Word Cloud from Frequency Table of Words
 #'
-#' Create a word cloud from a frequency table of words, and save to a PNG file.
-#' Visualizations are created using the
-#' [wordcloud2.js](https://wordcloud2-js.timdream.org/) JavaScript library.
+#' @description Create a word cloud from a frequency table of words, and save to a PNG file.
+#'   Requires R-packages \pkg{htmltools}, \pkg{htmlwidgets}, \pkg{magick}, \pkg{webshot2},
+#'   and \pkg{wordcloud2} are available.
+#'   System dependencies include the the following:
+#'   [ImageMagick](https://imagemagick.org/) for displaying the PNG image,
+#'   [OptiPNG](https://optipng.sourceforge.net/) for PNG file compression, and
+#'   [Chrome](https://www.google.com/chrome/)- or a Chromium-based browser
+#'   with support for the Chrome DevTools protocol.
+#'   Use [`find_chromate`][chromote::find_chrome] function to find the path to the Chrome browser.
 #'
 #' @param x 'data.frame'.
-#'   A frequency table of words that includes `"word"` and `"freq"` in each column.
+#'   A frequency table of words that includes "word" and "freq" in each column.
 #' @param max_words 'integer' number.
 #'   Maximum number of words to include in the word cloud.
 #' @param size 'numeric' number.
 #'   Font size, where the larger size indicates a bigger word.
 #' @param shape 'character' string.
 #'   Shape of the \dQuote{cloud} to draw.
-#'   Possible shapes include a `"circle"`, `"cardioid"`, `"diamond"`,
-#'   `"triangle-forward"`, `"triangle"`, `"pentagon"`, and `"star"`.
+#'   Possible shapes include a "circle", "cardioid", "diamond",
+#'   "triangle-forward", "triangle", "pentagon", and "star".
 #' @param ellipticity 'numeric' number.
 #'   Degree of \dQuote{flatness} of the shape to draw, a value between 0 and 1.
 #' @param ...
-#'   Additional arguments to be passed to the [`wordcloud2::wordcloud2`] function.
+#'   Additional arguments to be passed to the [`wordcloud2`][wordcloud2::wordcloud2] function.
 #' @param width 'integer' number.
 #'   Desired image width in pixels.
 #' @param output 'character' string.
 #'   Path to the output file, by default the word cloud is copied to a temporary file.
 #' @param display 'logical' flag.
 #'   Whether to display the saved PNG file in a graphics window.
-#'   Requires access to the \pkg{png} package.
+#'   Requires access to the \pkg{magick} package.
 #'
-#' @details PNG file compression requires the external program
-#'   [OptiPNG](https://optipng.sourceforge.net/), download and install.
-#'
-#' @return The word cloud plots in PNG format, and the path of the output file is returned.
+#' @return File path to the word cloud plot in PNG format.
 #'
 #' @author J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
-#'
-#' @keywords internal
 #'
 #' @export
 #'
 #' @examples
 #' \dontrun{
-#' make_word_cloud(wordcloud2::demoFreq, size = 1.5, display = TRUE)
+#'   file <- wordcloud2::demoFreq |>
+#'     make_wordcloud(size = 1.5, display = interactive())
+#'
+#'   unlink(file)
 #' }
 
-make_word_cloud <- function(x,
-                            max_words = 200L,
-                            size = 1,
-                            shape = "circle",
-                            ellipticity = 0.65,
-                            ...,
-                            width = 910L,
-                            output = NULL,
-                            display = FALSE) {
+make_wordcloud <- function(x,
+                           max_words = 200L,
+                           size = 1,
+                           shape = "circle",
+                           ellipticity = 0.65,
+                           ...,
+                           width = 910L,
+                           output = NULL,
+                           display = FALSE) {
 
-  # check package dependencies
+  # check packages
   pkgs <- c(
-    "grid",
     "htmltools",
     "htmlwidgets",
-    "png",
+    "magick",
     "utils",
     "webshot2",
     "wordcloud2"
   )
-  is <- !vapply(pkgs, requireNamespace, logical(1), quietly = TRUE)
-  if (any(is)) {
-    stop(
-      "A word cloud requires the following missing package(s): ",
-      paste(sQuote(pkgs[is]), collapse = ", "),
-      call. = FALSE
-    )
+  is <- vapply(pkgs, FUN = requireNamespace, FUN.VALUE = logical(1), quietly = TRUE)
+  if (any(!is)) {
+    txt <- pkgs[!is] |> sQuote(q = FALSE) |> paste(collapse = ", ")
+    stop("Word cloud requires missing package(s): ", txt, call. = FALSE)
   }
 
   # check arguments
@@ -117,7 +117,7 @@ make_word_cloud <- function(x,
 
   # configure to not display hover labels
   sty <- htmltools::HTML(".wcLabel {display: none;}")
-  tag <- htmltools::tags$head(htmltools::tags$style(sty))
+  tag <- htmltools::tags$style(sty) |> htmltools::tags$head()
   wc <- htmlwidgets::prependContent(wc, tag)
 
   # save word-cloud widget to a temporary HTML file
@@ -146,13 +146,16 @@ make_word_cloud <- function(x,
 
   # recompress png file
   suppressWarnings(
-    system2("optipng", c("-quiet", "-strip all", "-o7", shQuote(output)))
+    system2(
+      command = "optipng",
+      args = c("-quiet", "-strip all", "-o7", shQuote(output))
+    )
   )
 
   # display saved PNG in a graphics window
   if (display) {
-    grid::grid.raster(png::readPNG(output))
+    magick::image_read(output) |> plot()
   }
 
-  output
+  invisible(output)
 }